websocket message type handler cleanup
