+++ frontend/src/views/lobbyView.ts
@@ -1,6 +1,7 @@
 import { createDOMElement, useStateManager } from "mini-framework";
 import { connectWS } from "../websocket";
 import { navigateTo } from "../main";
+import { EnterLobbyServerMessage, TimerUpdates, wsEvent } from "../models/wsMessage";
 
 export let TimerCountDown = useStateManager("0")
 
@@ -32,5 +33,65 @@ export const lobbyView = () => {
     return HTML
 }
 
+export function addPlayerCount(eventData:EnterLobbyServerMessage){
+    const playerCountContainer = document.getElementById("playerCount")
+    const playerCount = eventData.playerCount
+
+    if (playerCountContainer) {
+        playerCountContainer.innerText = playerCount + " user(s) in the lobby right now"
+    }
+}
+
+export function twentySecondTimer(eventData:TimerUpdates){
+    const timerSeconds = eventData.seconds
+    const fristTimerText = document.getElementById("firstTimer")
+    let isHidden = true;
+
+    if (fristTimerText) {
+        isHidden = fristTimerText.classList.contains('hidden');
+    }
+
+    if (isHidden && fristTimerText) {
+        fristTimerText.classList.toggle('hidden');
+    }
+
+    if (fristTimerText) {
+        fristTimerText.innerText = "the countdown will begin in " + String(timerSeconds) + " seconds"
+    }
+}
+
+export function tenSecondTimer(eventData:TimerUpdates){
+    const timerseconds = eventData.seconds
+    const secondTimer = document.getElementById("secondTimer")
+
+    const previousTimer = document.getElementById("firstTimer")
+
+    let isSecondHidden = true;
+
+    if (previousTimer && !previousTimer.classList.contains("hidden")) {
+        previousTimer.classList.toggle("hidden")
+    }
+
+    //if there are less than 2 people in the lobby all of a sudden:
+    if (eventData.seconds === -1) {
+        if (secondTimer && !secondTimer.classList.contains('hidden')) {
+            secondTimer.classList.toggle('hidden');
+        }
+        return
+    }
+
+    if (secondTimer) {
+        isSecondHidden = secondTimer.classList.contains('hidden');
+    }
+
+    if (isSecondHidden && secondTimer) {
+        secondTimer.classList.toggle('hidden');
+    }
+
+    if (secondTimer) {
+        secondTimer.innerText = "the game will start in " + String(timerseconds) + " seconds"
+    }
+
+}
 
 // exmpl:  createDOMElement("div", {}, [])
